Deprecate string/table and move them to stdlib/utils/string stdlib/utils/table
Move deprecated files outside of stdlib and only pull in when making
release
