+++ deprecated/string.lua
@@ -1,0 +1,2 @@
+--deprecated: update require to stdlib/utils/string
+require 'stdlib/util/string'
