+++ deprecated/table.lua
@@ -1,0 +1,2 @@
+--deprecated: update require to stdlib/utils/table
+require 'stdlib/utils/table'
