+++ spec/entity/inventory_spec.lua
@@ -1,5 +1,5 @@
 local Inventory = require 'stdlib/entity/inventory'
-require 'stdlib/table'
+require 'stdlib/utils/table'
 
 describe('Inventory Spec', function()
     local function make_get_contents(inv)
