+++ spec/event/gui_spec.lua
@@ -155,7 +155,7 @@ describe('Gui', function()
 
     it('.dispatch should print an error to connected players if a handler throws an error', function()
         _G.game.players = { { name = 'test_player', valid = true, connected = true, print = function() end } }
-        require('stdlib/table')
+        require('stdlib/utils/table')
         _G.game.connected_players = table.filter(_G.game.players, function(p) return p.connected end)
 
         local s = spy.on(_G.game, "print")
