+++ spec/utils/string_spec.lua
@@ -1,4 +1,4 @@
-require 'stdlib/string'
+require 'stdlib/utils/string'
 
 describe('String Spec', function()
     describe('string.trim', function()
