+++ stdlib/event/force.lua
@@ -8,7 +8,7 @@
 -- -- The fist time this is required it will register force creation events
 
 require('stdlib/event/event')
-require('stdlib/table')
+require('stdlib/utils/table')
 local Game = require 'stdlib/game'
 
 local Force = {}
